Add kontrol of empty content. Also additional styling

+++ src/js/main.ts
@@ -26,40 +26,55 @@ function getCourses() {
         .then(data => {
             data.forEach(course => {
                 coursesEl.innerHTML +=
-                    `<div class="course flex-container">
+                    `<div class="course">
                 <div class="courseEl">${course.course_id} </div>
                 <div class="courseEl">${course.name} </div>
                 <div class="courseEl">${course.progression} </div>
                 <div class="courseEl"><a href="${course.course_syllabus}" target="_blank">Länk</a></div>
-                <button class="delete-button" id="${course.id}" onClick="deleteCourse(${course.id})">Ta bort <i class="fas fa-trash-alt"></i></button>
-                <button class="edit-button" onClick="editCourse('${course.id}', '${course.course_id}', '${course.name}', '${course.progression}', '${course.course_syllabus}')">Redigera <i class="far fa-edit"></i></button>
+                <button class="delete-button" id="${course.id}" onClick="deleteCourse(${course.id})"><i class="fas fa-trash-alt"></i></button>
+                <button class="edit-button" onClick="editCourse('${course.id}', '${course.course_id}', '${course.name}', '${course.progression}', '${course.course_syllabus}')"><i class="far fa-edit"></i></button>
             </div>`;
             })
         });
 };
 
 function deleteCourse(id: number) {
-    fetch("http://localhost:8080/webbtjanst/rest?id=" + id, {
-        method: "DELETE",
-    })
-        .then(response => response.json())
-        .then(data => {
-            getCourses();
+    // Begär bekräftelse
+    if(confirm("Är du säker på att du vill ta bort den här kursen?")) {
+        // Skickar sen till webbtjänsten samt skriver ut kurserna på nytt
+        fetch("http://localhost:8080/webbtjanst/rest?id=" + id, {
+            method: "DELETE",
         })
-        .catch(error => {
-            console.log("Error: ", error);
-        });
+            .then(response => response.json())
+            .then(data => {
+                getCourses();
+            })
+            .catch(error => {
+                console.log("Error: ", error);
+            });
+    } else {
+        return false; // Om användaren inte vill radera händer ingenting
+    }
 };
 
 function addCourse() {
-    let code = codeInput.value;
-    let name = nameInput.value;
-    let prog = progInput.value;
-    let syllabus = syllabusInput.value;
+    let code: string = codeInput.value;
+    let name: string = nameInput.value;
+    let prog: string = progInput.value;
+    let syllabus: string = syllabusInput.value;
+
+    // Kontrollerar värden
+
+    if(code == "" || name == "" || prog == "" || syllabus == ""){
+        alert("Alla fält måste fyllas i!");
+        return false;
+    };
 
     let course = { "course_id": code, "name": name, "progression": prog, "course_syllabus": syllabus };
 
     console.log(course);
+
+    
 
     fetch("http://localhost:8080/webbtjanst/rest", {
         method: "POST",
@@ -83,16 +98,16 @@ function editCourse(id: number, courseId: string, courseName: string, courseProg
         <form>
             <h3>Redigera kurs</h3>
             <label for="code-edit">Kurskod</label>
-            <input type="text" name="code-edit" id="code-edit" value="${courseId}">
+            <input type="text" name="code-edit" id="code-edit" value="${courseId}" required>
 
             <label for="name-edit">Kursnamn</label>
-            <input type="text" name="name-edit" id="name-edit" value="${courseName}">
+            <input type="text" name="name-edit" id="name-edit" value="${courseName}" required>
 
             <label for="prog-edit">Progression</label>
-            <input type="text" name="prog-edit" id="prog-edit" value="${courseProg}">
+            <input type="text" name="prog-edit" id="prog-edit" value="${courseProg}" required>
 
             <label for="syllabus-edit">Kursplan(länk)</label>
-            <input type="text" name="syllabus-edit" id="syllabus-edit" value="${courseSyllabus}">
+            <input type="text" name="syllabus-edit" id="syllabus-edit" value="${courseSyllabus}" required>
 
             <div class="flex-container">
                 <button id="abort" onClick="abortEdit()">Avbryt</button>
@@ -126,6 +141,13 @@ function updateCourse(id: number) {
     name.toString();
     prog.toString();
     syllabus.toString();
+
+    // Kontrollerar värden
+
+    if(code == "" || name == "" || prog == "" || syllabus == ""){
+        alert("Alla fält måste fyllas i!");
+        return false;
+    };
 
     let course = { "course_id": code, "name": name, "progression": prog, "course_syllabus": syllabus };
